Don't change the media controls visibility in MediaControls::reset()
reset() is called from a number of places with the expectation of
putting the controls into the state they would be if created anew.

One such place is via HTMLMediaElement::preDispatchEventHandler, where
the controls are reset when entering or exiting fullscreen. This has a
very strange side-effect, namely that dispatching a synthetic
webkitfullscreenchange event on a video will cause the media controls to
become visible.

It would be rather messy to write a layout test for this, and it would
be slow to run, so instead add unit tests to test this more directly.

R=fs@opera.com

Review URL: https://codereview.chromium.org/1319433003

git-svn-id: svn://svn.chromium.org/blink/trunk@201126 bbb929c8-8fbe-4397-9dbb-9b2b20218538