Fallback to root layer if hit-testing does not hit anything in iframe
Currently we have a hit-testing fallback that if nothing is hit,
consider the root layer (document element) is hit. The fallback is
only for main document, but not iframe. It is not unreasonable that
if the hit point is within the iframe's bound, it should be considered
to hit the iframe's document if it does not hit any other element.
This CL does this modification to the fallback rule.

BUG=508474

Review URL: https://codereview.chromium.org/1289753006

git-svn-id: svn://svn.chromium.org/blink/trunk@201188 bbb929c8-8fbe-4397-9dbb-9b2b20218538