Fix &nbsp; fallback glyph lookup
On Japanese locale Android which has Motoya Maruberi
as the system default font (up until 5.1), the font
fallback path mistakenly picked up ".notdef" glyph
for &nbsp; (U+00A0).

This happened because in the fallback path it still
ended up in looking up non-existent glyph in the same
font file and fell back to the "missingGlyphData".

This is because Blink looks up the glyph in
normalized codepoint (for U+00A0, normalized to U+0020)
but still it tries to look up the original codepoint.

This bug was exposed by the following change, to make
not to normalize &nbsp; to space before looking up glyph.
https://codereview.chromium.org/705163003

BUG=454108

Review URL: https://codereview.chromium.org/1317833003

git-svn-id: svn://svn.chromium.org/blink/trunk@201348 bbb929c8-8fbe-4397-9dbb-9b2b20218538