Make PartitionAllocDeathTest.GuardPages pass on Win64.
On Windows, the allocation granularity is 64kB instead of 4kB elsewhere,
so if a multiple of 4kB data is requested from partitionAlloc() there
is some free space at the end.  On 32-bit 4kB following the allocation
is mapped as inaccessible, but on 64-bit this isn't done since ASLR
should cover writes that are far enough out-of-bounds.  So make sure
GuardPages writes outside of the allocated memory, and not into the
free space at the end of the allocation.

BUG=524308

Review URL: https://codereview.chromium.org/1314843008

git-svn-id: svn://svn.chromium.org/blink/trunk@201453 bbb929c8-8fbe-4397-9dbb-9b2b20218538