Make XPathFunction map use non-static table of FunctionMapping
The functionMap is already a global static variable. The createFunction has a check
which makes the method createFunction being called only once. Hence there is no need
of this function table to be a static table.

This reduces the bss size on linux by 656 bytes

R=haraken

Review URL: https://codereview.chromium.org/1324133002

git-svn-id: svn://svn.chromium.org/blink/trunk@201638 bbb929c8-8fbe-4397-9dbb-9b2b20218538