Oilpan: Fix a cycle caused by SVGPropertyTearOffBase::m_contextElement
For the reason described in the bug, using a Member for SVGPropertyTearOffBase::m_contextElement
causes a cycle. To break the cycle, this CL changes it to a raw pointer.
The raw pointer is safe because the context element is guaranteed to be kept alive
by a V8 wrapper of the SVG tear off object.

BUG=528275

Review URL: https://codereview.chromium.org/1330673002

git-svn-id: svn://svn.chromium.org/blink/trunk@201796 bbb929c8-8fbe-4397-9dbb-9b2b20218538