Simplify handling of initialized AudioContext
AbstractAudioContext included a variable m_isInitialized.  This was basically set to true or false depending on whether the destination node was initialized or not, essentially duplicating the information in the destination node.

Remove m_isInitialized from AbstractAudioContext, and use the m_isInitialized slot from the destination node to determine if the audio context is ready for processing audio.  m_isInitialized is then no longer written to by the main thread, and all initialization (and uninitialization) is handled by the destination node.

Renamed AbstractAudioContext::isInitialized() to isDestinationInitialized() to make it clearer what is being checked.

BUG=527733
TEST=test case no longer triggers TSAN

Review URL: https://codereview.chromium.org/1331953003

git-svn-id: svn://svn.chromium.org/blink/trunk@202217 bbb929c8-8fbe-4397-9dbb-9b2b20218538