Oilpan: Remove reserveInitialCapacity from DistributedNodes' constructor
In short:

This CL is necessary to fix performance regressions observed in Oilpan.
This CL doesn't affect performance of non-Oilpan.

In details:

The reserveInitialCapacity(32) was introduced in https://chromiumcodereview.appspot.com/23766020.
As far as I read the CL description, it seems that the reserveInitialCapacity(32) was introduced
just to prevent a regression in micro-benchmarks without doing serious performance measurement
in realistic scenarios. The reserveInitialCapacity(32) may be working well in non-Oilpan,
but at the very least, it is causing a regression in Oilpan.

With Oilpan enabled, allocating the vector buffer (of size sizeof(Node)*32) adds a substantial
amount of pressures on the heap and thus increases a GC frequency. The regression
caused by the GCs is larger than the improvement caused by the pre-allocation.
Overall, it is better to not call reserveInitialCapacity(32) in Oilpan.
For example, document.createElement("select") is 2x slower in Oilpan than in non-Oilpan.

Thus this CL removes reserveInitialCapacity(32) from DistributedNodes' constructor.

This CL removes an unused method from DistributedNodes.

BUG=

Review URL: https://codereview.chromium.org/1333813002

git-svn-id: svn://svn.chromium.org/blink/trunk@202244 bbb929c8-8fbe-4397-9dbb-9b2b20218538